--- src/components/FavoriteProducts.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const FavoriteProducts = () => {
-  return (
-    <div>
-      FavoriteProducts
-    </div>
-  )
-}
-
-export default FavoriteProducts
