+++ src/components/PaginationLink.tsx
@@ -17,8 +17,6 @@ const PaginationLink = ({
 }: PaginationLinkProps) => {
 
   const params = useSearchParams();
-  const limit = PRODUCTS_PER_PAGE;
-  const skip = page ? (Number(page) - 1) * limit : 0;
 
   let currentQuery = {};
